Moved responsabilities arround and made a level new format level generator. We shall now put an adaptor to the old classes so we support the new level generation.
